+++ README.md
@@ -14,17 +14,17 @@ https://ovenapp.io/project/Ckt29wbbfP9DgHDEQzQV8HJF6BG0Vc5O#E9c6C
         - [X] 사용자의 여행일지에 추가됨.
 - **회원기능**
     - [X] 회원가입, 로그인
-    - [ ] 온라인 상태정보 제공, 껐다 킬 수 있다.
+    - [X] 온라인 상태정보 제공, 껐다 킬 수 있다.
     - [X] 프로필 이미지
 - **주변 사용자와 소통**
-    - [ ] 현재 위치 반경기반 주변 사용자 검색
-    - [ ] 주변 사용자와 1:1대화 가능함
-    - [ ] 온라인 상태가 꺼진 사용자는 검색대상에서 제외
+    - [X] 현재 위치 반경기반 주변 사용자 검색
+    - [X] 주변 사용자와 1:1대화 가능함
+    - [X] 온라인 상태가 꺼진 사용자는 검색대상에서 제외
 - **랭킹 시스템**
     - [X] 월 별 인기 여행지를 n개 나열해서 보여줌
     - [X] 일, 주, 월 인기 피드 조회
-    - [ ] 년별 인기 여행일지 조회
-    - [ ] 일, 주 별 인기 사용자 보여줌
+    - [X] 년별 인기 여행일지 조회
+    - [X] 일, 주 별 인기 사용자 보여줌
 - **검색기능**
     - [X] 사용자 닉네임으로 사용자 검색
     - [X] 태그로 피드 검색
