+++ batch/build.gradle
@@ -1,0 +1,25 @@
+plugins {
+    id 'java'
+}
+
+bootJar {
+    mainClass = 'com.stoury.ChatStouryApplication'
+}
+
+group 'com'
+version '0.0.1-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+
+    implementation rootProject
+}
+
+test {
+    useJUnitPlatform()
+}
