+++ batch/src/main/java/com/stoury/batch/BatchPopularSpotsConfig.java
