+++ build.gradle
@@ -18,7 +18,7 @@ allprojects {
     apply plugin: 'groovy'
     apply plugin: 'org.springframework.boot'
     apply plugin: 'io.spring.dependency-management'
-    apply plugin:  "org.asciidoctor.jvm.convert"
+    apply plugin: "org.asciidoctor.jvm.convert"
 
     configurations {
         asciidoctorExt
@@ -40,6 +40,8 @@ allprojects {
         runtimeOnly 'org.mariadb.jdbc:mariadb-java-client'
         annotationProcessor 'org.projectlombok:lombok'
         implementation 'org.springframework.boot:spring-boot-starter-validation'
+        implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+        implementation 'org.springframework.session:spring-session-data-redis'
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
         testImplementation 'org.springframework.security:spring-security-test'
         testImplementation 'org.springframework.restdocs:spring-restdocs-mockmvc'
@@ -67,7 +69,7 @@ allprojects {
 
     bootJar {
         dependsOn asciidoctor
-        from ("${asciidoctor.outputDir}/html5") {
+        from("${asciidoctor.outputDir}/html5") {
             into 'static/docs'
         }
     }
@@ -99,19 +101,21 @@ allprojects {
     }
 }
 
-project(':chats'){
+project(':chats') {
     dependencies {
         implementation 'org.springframework.kafka:spring-kafka'
         testImplementation 'org.springframework.kafka:spring-kafka-test'
     }
 }
 
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
-    implementation 'org.springframework.boot:spring-boot-starter-batch'
-    implementation 'org.springframework.session:spring-session-data-redis'
-    testImplementation 'org.springframework.batch:spring-batch-test'
+project(':batch') {
+    dependencies {
+        implementation 'org.springframework.boot:spring-boot-starter-batch'
+        testImplementation 'org.springframework.batch:spring-batch-test'
+    }
+}
 
+dependencies {
     // 구글 지오코딩 api
     implementation 'com.google.maps:google-maps-services:2.2.0'
 }
